* extract data classes into separate files * move api-specific requests/responses out of models package
