+++ app/src/main/java/com/reach5/identity/sdk/demo/JavaMainActivity.java
@@ -13,9 +13,9 @@ import androidx.appcompat.app.AppCompatActivity;
 import com.reach5.identity.sdk.core.JavaReachFive;
 import com.reach5.identity.sdk.core.Provider;
 import com.reach5.identity.sdk.core.models.AuthToken;
-import com.reach5.identity.sdk.core.models.SdkConfig;
+import com.reach5.identity.sdk.core.SdkConfig;
 import com.reach5.identity.sdk.core.models.OpenIdUser;
-import com.reach5.identity.sdk.core.models.requests.ProfileSignupRequest;
+import com.reach5.identity.sdk.core.models.ProfileSignupRequest;
 import com.reach5.identity.sdk.google.GoogleProvider;
 import com.reach5.identity.sdk.webview.WebViewProvider;
 import io.github.cdimascio.dotenv.Dotenv;
