+++ sdk-core/src/main/java/com/reach5/identity/sdk/core/JavaReachFive.kt
@@ -3,11 +3,10 @@ package com.reach5.identity.sdk.core
 import android.app.Activity
 import android.content.Intent
 import com.reach5.identity.sdk.core.models.*
-import com.reach5.identity.sdk.core.models.requests.ProfileSignupRequest
-import com.reach5.identity.sdk.core.models.requests.UpdatePasswordRequest
+import com.reach5.identity.sdk.core.models.ProfileSignupRequest
+import com.reach5.identity.sdk.core.models.UpdatePasswordRequest
+import com.reach5.identity.sdk.core.models.ReachFiveError
 import com.reach5.identity.sdk.core.utils.Callback
-import com.reach5.identity.sdk.core.utils.Failure
-import com.reach5.identity.sdk.core.utils.SuccessWithNoContent
 
 class JavaReachFive(activity: Activity, sdkConfig: SdkConfig, providersCreators: List<ProviderCreator>) {
     private val reach5 = ReachFive(activity, sdkConfig, providersCreators)
