+++ sdk-core/src/main/java/com/reach5/identity/sdk/core/Provider.kt
@@ -6,7 +6,6 @@ import com.reach5.identity.sdk.core.api.ReachFiveApi
 import com.reach5.identity.sdk.core.models.AuthToken
 import com.reach5.identity.sdk.core.models.ProviderConfig
 import com.reach5.identity.sdk.core.models.ReachFiveError
-import com.reach5.identity.sdk.core.models.SdkConfig
 import com.reach5.identity.sdk.core.utils.Failure
 import com.reach5.identity.sdk.core.utils.Success
