+++ sdk-core/src/main/java/com/reach5/identity/sdk/core/SdkConfig.kt
@@ -1,4 +1,4 @@
-package com.reach5.identity.sdk.core.models
+package com.reach5.identity.sdk.core
 
 import android.os.Parcelable
 import kotlinx.android.parcel.Parcelize
