+++ sdk-core/src/main/java/com/reach5/identity/sdk/core/api/requests/SignupRequest.kt
@@ -1,7 +1,8 @@
-package com.reach5.identity.sdk.core.models.requests
+package com.reach5.identity.sdk.core.api.requests
 
 import android.os.Parcelable
 import com.google.gson.annotations.SerializedName
+import com.reach5.identity.sdk.core.models.ProfileSignupRequest
 import kotlinx.android.parcel.Parcelize
 
 @Parcelize
