+++ sdk-core/src/main/java/com/reach5/identity/sdk/core/api/requests/UpdatePhoneNumberRequest.kt
@@ -1,4 +1,4 @@
-package com.reach5.identity.sdk.core.models.requests
+package com.reach5.identity.sdk.core.api.requests
 
 import android.os.Parcelable
 import com.google.gson.annotations.SerializedName
