+++ sdk-core/src/main/java/com/reach5/identity/sdk/core/models/ProfileSignupRequest.kt
@@ -1,9 +1,7 @@
-package com.reach5.identity.sdk.core.models.requests
+package com.reach5.identity.sdk.core.models
 
 import android.os.Parcelable
 import com.google.gson.annotations.SerializedName
-import com.reach5.identity.sdk.core.models.Consent
-import com.reach5.identity.sdk.core.models.ProfileAddress
 import kotlinx.android.parcel.Parcelize
 import kotlinx.android.parcel.RawValue
