+++ sdk-facebook/src/main/java/com/reach5/identity/sdk/facebook/FacebookProvider.kt
@@ -10,10 +10,12 @@ import com.facebook.login.LoginManager
 import com.facebook.login.LoginResult
 import com.reach5.identity.sdk.core.Provider
 import com.reach5.identity.sdk.core.ProviderCreator
+import com.reach5.identity.sdk.core.SdkConfig
 import com.reach5.identity.sdk.core.api.ReachFiveApi
 import com.reach5.identity.sdk.core.api.ReachFiveApiCallback
 import com.reach5.identity.sdk.core.models.*
-import com.reach5.identity.sdk.core.models.requests.LoginProviderRequest
+import com.reach5.identity.sdk.core.api.requests.LoginProviderRequest
+import com.reach5.identity.sdk.core.models.ReachFiveError
 import com.reach5.identity.sdk.core.utils.Failure
 import com.reach5.identity.sdk.core.utils.Success
