+++ sdk-google/src/main/java/com/reach5/identity/sdk/google/GoogleProvider.kt
@@ -15,10 +15,12 @@ import com.google.android.gms.common.api.GoogleApiClient.ConnectionCallbacks
 import com.google.android.gms.common.api.Scope
 import com.reach5.identity.sdk.core.Provider
 import com.reach5.identity.sdk.core.ProviderCreator
+import com.reach5.identity.sdk.core.SdkConfig
 import com.reach5.identity.sdk.core.api.ReachFiveApi
 import com.reach5.identity.sdk.core.api.ReachFiveApiCallback
 import com.reach5.identity.sdk.core.models.*
-import com.reach5.identity.sdk.core.models.requests.LoginProviderRequest
+import com.reach5.identity.sdk.core.api.requests.LoginProviderRequest
+import com.reach5.identity.sdk.core.models.ReachFiveError
 import com.reach5.identity.sdk.core.utils.Failure
 import com.reach5.identity.sdk.core.utils.Success
