+++ sdk-webview/src/main/java/com/reach5/identity/sdk/webview/WebViewProviderConfig.kt
@@ -3,7 +3,7 @@ package com.reach5.identity.sdk.webview
 import android.os.Parcel
 import android.os.Parcelable
 import com.reach5.identity.sdk.core.models.ProviderConfig
-import com.reach5.identity.sdk.core.models.SdkConfig
+import com.reach5.identity.sdk.core.SdkConfig
 import com.reach5.identity.sdk.core.models.SdkInfos
 import com.reach5.identity.sdk.core.utils.Pkce
 import java.net.URLEncoder
